requirements.txt
python-telegram-bot==20.7
moviepy==1.0.3
Pillow==10.2.0
